requirements.txt

selenium
boto3
requests
python-dotenv
